+++ lib/apress/yandex_market/presenters.rb
@@ -1,6 +1,7 @@
 module Apress
   module YandexMarket
     module Presenters
+      autoload :Base, 'apress/yandex_market/presenters/base'
       autoload :Model, 'apress/yandex_market/presenters/model'
     end
   end
